requirements.txt
datasets==1.6.2
matplotlib==3.7.5
networkx==3.1
numpy==1.22.3
openpyxl==3.1.5
PubChemPy==1.0.4
pyarrow==17.0.0
pytorch-fast-transformers==0.4.0
pytorch-lightning==1.1.5
rdkit==2023.9.4
regex==2024.5.15
scikit-learn==0.24.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
SmilesPE==0.0.3
tensorboard==2.14.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchdata==0.8.0
torchvision==0.8.2+cu110
umap-learn==0.5.6

